fix(ui): #2051: remove redundant summing logic
